Correcciones

+++ Segunda-entrega/src/dao/dbManager/products.dao.js
@@ -3,24 +3,25 @@ import { productModel } from "../../models/products.model.js";
 class ProductDao {
   async getAllProducts(limit, page, sort, filter) {
     const query = {};
-    if (filter != undefined) {
-      if (
-        filter.split(":")[0] == "title" ||
-        filter.split(":")[0] == "category" ||
-        filter.split(":")[0] == "description"
-      ) {
-        query[filter.split(":")[0]] = {
-          $regex: new RegExp(filter.split(":")[1], "i"),
-        };
-      } else if (
-        filter.split(":")[0] == "_id" ||
-        filter.split(":")[0] == "price" ||
-        filter.split(":")[0] == "code" ||
-        filter.split(":")[0] == "status"
-      ) {
-        query[filter.split(":")[0]] = filter.split(":")[1];
+    if (filter !== undefined) {
+      const [field, value] = filter.split(":");
+      const validFields = [
+        "title",
+        "category",
+        "description",
+        "_id",
+        "price",
+        "code",
+        "status",
+      ];
+      if (validFields.includes(field)) {
+        if (["title", "category", "description"].includes(field)) {
+          query[field] = { $regex: new RegExp(value, "i") };
+        } else {
+          query[field] = value;
+        }
       } else {
-        throw Error("El campo que desea filtrar no existe.");
+        throw new Error("El campo que desea filtrar no existe.");
       }
     }
     const orderedQuery = productModel
@@ -42,8 +43,8 @@ class ProductDao {
         hasPrevPage: products.hasPrevPage,
         hasNextPage: products.hasNextPage,
       };
-      newObject.nextLink = `/products?page=${newObject.nextPage}&&limit=${limit}`
-      newObject.prevLink = `/products?page=${newObject.prevPage}&&limit=${limit}`
+      newObject.nextLink = `/products?page=${newObject.nextPage}&&limit=${limit}`;
+      newObject.prevLink = `/products?page=${newObject.prevPage}&&limit=${limit}`;
       return newObject;
     } catch (e) {
       throw Error(e.message);
@@ -55,14 +56,56 @@ class ProductDao {
   }
 
   async createProduct(product) {
+    const fieldsValidation = this.validateFields(product);
+    if (fieldsValidation) {
+      throw Error(
+        "Para ingresar un nuevo producto, es necesario completar los campos obligatorios."
+      );
+    }
+    const codeValidation = await this.validateCode(product);
+    if (codeValidation) {
+      throw Error(
+        `El 'code' del producto ingresado, ya existe en el gestionador de Productos.`
+      );
+    }
     return await productModel.create(product);
   }
 
+  validateFields(product) {
+    const arrayProduct = [
+      product.title,
+      product.description,
+      product.price,
+      product.code,
+      product.category,
+      product.stock,
+    ];
+    const required =
+      arrayProduct.includes(undefined) || arrayProduct.includes("");
+    return required;
+  }
+
+  async validateCode(product) {
+    const products = await productModel.find();
+    const validation = products.some((el) => el.code === product.code);
+    return validation;
+  }
+
   async updateProduct(id, product) {
+    const codeValidation = await this.validateCode(product);
+    if (codeValidation) {
+      throw Error(
+        `El 'code' del producto ingresado, ya existe en el gestionador de Productos.`
+      );
+    }
     return await productModel.findByIdAndUpdate(id, product);
   }
 
   async deleteProduct(id) {
+    const product = await productModel.findById(id);
+    if (!product) {
+      throw Error(`El producto con id ${id} no existe.`);
+    }
     return await productModel.findByIdAndDelete(id);
   }
 }
